Fixed APN Enhanced format error detection
The APNS Enhanced format returns a response if a notification failed,
but then it also closes the connection.  This allows for a possibility
of sending one notification, then sending a second one immediately
after, and then getting notified that the first one failed.  What
happens in this case is that the second one looks like it's been sent to
us, but apple closes the stream after the second one was sent, without
acknowledging that it failed or wasn't sent.  This means we have to wait
for possible errors from each message on the connection before sending
the next.  This sucks, and means an added 250 ms delay to sending each
message.  This also means that to speed up message delivery, more
Notification Channels are needed to scale across connections/threads.
Unavoidable if the Enhanced format is to be used (which provides the
ability to expire notifications as well as error reporting).